apps/hypnos/scripts/requirements.txt
selenium
pymongo
Flask==2.1.2
python-dotenv